refactor: linesteps로 line 만드는 메소드 이동
